refactor: 폴더분리 및 클래스명 변경
