+++ src/main/java/racingcar/model/car/factory/CarFactory.java
@@ -1,4 +1,7 @@
-package racingcar.model.car;
+package racingcar.model.car.factory;
+
+import racingcar.model.car.Car;
+import racingcar.model.car.strategy.MoveStrategy;
 
 public class CarFactory {
     public static Car createCar(
