+++ src/main/java/racingcar/model/car/strategy/MoveStrategy.java
@@ -1,4 +1,4 @@
-package racingcar.model.car;
+package racingcar.model.car.strategy;
 
 public interface MoveStrategy {
     boolean canMove();
