+++ src/test/java/racingcar/model/car/CarTest.java
@@ -5,6 +5,7 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.model.car.strategy.AlwaysMoveStrategy;
 
 public class CarTest {
